Fix debugger stepping into Main method
This fixes a long standing pet peeve of mine: F11 into the
`__managed_Main` method doesn't work and steps over the method instead.
This adds infrastructure so that we can do `DebuggerGuidedStepThrough`
on generated methods.

A slight annoyance is that for UTC, the FEEFEE sequence point needs to
be on a NOP that precedes the call, and that scheme doesn't work for
RyuJIT. But UTC will need to fix that anyway because we won't be
rewriting IL for them in the future and no-NOP is how
DebuggerGuidedStepThrough methods are in C# when used from within the
BCL.