Fix type loader handling of __Canon in CoreRT
On Project N, `__Canon` is backed by an actual type in System.Private.CoreLib. In the new type system, we no longer need each type to have a token and we chose `__Canon` to be a synthetic type created out of thin air.

This poses a small problem when trying to refer to it from C#.

In this change:
* I'm making us not include `__Canon.cs` on CoreRT anymore.
* I'm updating the single method mode command line parser to parse `__Canon` specially. We were already special casing it by rewriting the type tree anyway (parse as an ECMA type and then replace). (This is what Project X uses in the bridging code too.)
* I'm making `GetCanonType` an intrinsic (which has access to the `__Canon` TypeDesc)
* Putting some code under ifdef that doesn't apply on CoreRT

[tfs-changeset: 1672488]