Unify pseudo-attribute code between NativeFormat and Ecma
This was following the old layering rules - resulting
in pointless duplication of logic between Native
and Ecma.

As a bonus, this also fixes an edge-case bug.
PreserveSigAttribute is not valid on constructors
and isn't synthesized by the full framework even
if you force that bit on. Project N now matches
that behavior.