Fix incorrect arm ifdef (#4585)
Everywhere in the native code except for two places, we use
`_TARGET_ARM_` as a define to identify compilation for ARM32.
But at those two places, we use `TARGET_ARM`. This is incorrect since
`TARGET_ARM` is not defined anywhere for CoreRT builds.
This change replaces those two usages by the `_TARGET_ARM_`.