Compat: Disallow boxing return type conversion by CreateDelegate()
On CoreCLR,

   MethodInfo m = ..."Moo"..;
   m.CreateDelegate(typeof(Func<Object>))

   static int Moo() => 7;

throws an ArgumentException regarding a mismatching signature.

On Project N, the CreateDelegate() succeeds. But the celebration
turns into tears as soon as you invoke the delegate
as the hoped for boxing never occurs. Instead you get an unboxed
integer stored into a variable that expects an object reference.