Introduce NetCore 1.0 compat hack for GetTypeInfo() (#4711)
We encountered a scenario with F# where a user-extended
Type object was passed to another component that used
GetTypeInfo() on the object.

The user-extended Type object seemingly had no way
to implement IReflectableType.GetTypeInfo() because
there's no protected constructor on TypeInfo. But luckily,
TypeDelegator already extends TypeInfo so returning an instance
of that was enough to make the other component happy.

This is a cheap insurance to avoid getting other
reports of this.