Bug fixes for dynamic reflection invoke using the call converter:
1) Rassing InvokeUtils.DynamicInvokeParamType.Ref flag to InvokeUtils.DynamicInvokeParamHelperCore() only when the argument is explicitly passed byref (method signature has 'byref').

2) When the argument type of the callee was a reference type, we were using typeof(object) as the arg type, and calling InvokeUtils.DynamicInvokeParamHelperCore (which checks for argument types). This is incorrect, and we should use the exact typehandle of the argument type when calling InvokeUtils.DynamicInvokeParamHelperCore.
To fix this, I changed all the call converter APIs to always return the TypeHandle of the argument/return type, and not just in the cases of valuetypes. There is no reason to return default(TypeHandle) for reference types. There is an API called IsValueType on TypeHandle. If we need to check if a th is valuetype, just call the IsValueType API instead of .IsNull (which makes a lot more sense really).

3) Reflection invokers emulated by the calling convention converter, and returning void, need to return a 'null' object back to the caller (the signature of the invoker has return type of object. The statically compiled versions of these invokers return null in the "return void" scenarios)

[tfs-changeset: 1679620]