Adding handling for satellite assemblies in ProjectX to enable ResourceManager scenarios to run correctly.
The idea behind these changes is to enable the list of satellite assemblies to be loaded as ModuleDescs inside the compiler, so that they can be naturally processed for resources by the ResourceDataNode.

Currently in ProjectN the IL2IL toolchain extracts resources and writes them out to .rsrc files, with .index.txt files, that NUTC reads and dumps into the BlobIdResourceIndex native format blob.
These changes will prevent the .rsrc and .txt files from being emitted in purenative compilation mode, and instead we'll pipe the list of satellite assemblies directly to the ILCompiler, so that they can be loaded as ModuleDescs and enumerated by the PrecomputedMetadataManager.

Note: we'll need to also do this for the CoreRT scenario. These changes only enable the PX scenario.

[tfs-changeset: 1680346]