Define CORERT for ProjectN builds
CoreRT is component of ProjectN (aka .NET Native for UWP), but our current defines do not reflect it because of CORERT is not defined in ProjectN builds. This creates confusion for folks outside the core runtime team, and conflicts with our documentation. For example, https://github.com/dotnet/corert/blob/master/Documentation/intro-to-corert.md says "CoreRT is the .NET Core runtime that is optimized for AOT scenarios, which .NET Native targets". It does not say "CoreRT is the .NET Core runtime optimized for AOT scenarios that uses RyuJIT backend" or anything like that.

- Before this change:
     PROJECTN: defined in closed source ProjectN builds
     CORERT: defined in open source CoreRT builds
     CORECLR: defined in CoreCLR builds

- After this change:
     PROJECTN: defined in closed source ProjectN builds (same)
     CORERT: defined in both open source CoreRT builds and closed source ProjectN builds (different)
     CORECLR: defined in CoreCLR builds (same)

[tfs-changeset: 1680901]