Fix two issues in managed number formatting (dotnet/coreclr#15105)
* Fix ValueStringBuilder to return correct array to pool

When growing the builder, it's erroneously returning the new array rather than the old one.

* Fix number formatting to not share static buffer

Was missing [ThreadStatic] on a static, such that parallel formatting of numbers might end up sharing the same buffer.  Oops.  But instead of just adding [ThreadStatic], I've switched to starting with a relatively small buffer on the stack.  It should suffice for the majority of performed formatting, and the ValueStringBuilder will grow for longer needs.

* Work around C# compiler warning about stack space being exposed

Signed-off-by: dotnet-bot <dotnet-bot@microsoft.com>