Implement efficient Array.IndexOf (#4977)
* Implement efficient Array.IndexOf

This is the remainder of #4935.

We special case a couple methods on Array to avoid allocating a full EqualityComparer per each array type. This saves 22 kB on hello world (comparers for a bunch of primitive types and enums).

Same as in #4935 - we restrict the optimization to non-shared value types. On Project N we would also optimize shared valuetypes, but these are rare enough that it's probably not worth building the infrastructure for.

Fixes #763.