Avoid rooting field reflection from InvokeUtils
InvokeUtils was calling the general purpose Enum.ToObject API instead of doing the low level things it's supposed to do. Enum.ToObject is a general purpose reflection API that will go and enumerate literal fields on the type to determine the underlying type. Also, query base type hierarchy in reflection terms to make sure the type in question is an enum.

This is a general purpose perf improvement.

[tfs-changeset: 1681991]