Fix up stack trace formatter to support general purpose metadata (#5016)
The stack trace formatter was written to support the metadata format emitted by the stack trace metadata emitter, but it turns out it doesn't support the more general purpose constructs.

The general purpose signature for a method on a generic type that takes T as a parameter is `ReturnType Foo(!0)`. The formatter needs to format the parameter `!0` as `T` (i.e. we need to do substitution for formals).

The code is less than ideal, mostly because it also needs to support metadata constructs that would be illegal in ECMA-335, but are relatively compact when it comes to binary representation, so we use them in the stack trace emitter.