Avoid provider-related costs for default integer formatting (#15272)
One of the biggest costs when invoking Int32.ToString() and similar methods is in fetching the current culture, accessing its number information, etc.  And if a culture is provided explicitly, there are still costs related to casts.  But when the default format is used for non-negative integer values, the number format information isn't used at all, and we can skip the associated costs.  This improves formatting perf in such cases by upwards of ~25%.

Signed-off-by: dotnet-bot <dotnet-bot@microsoft.com>