Unix: Remove MaxPath, MaxName (dotnet/coreclr#15229)
* Unix: Remove MaxPath, MaxName

PATH_MAX is not an upper bound on path lengths on Unix. system function return
values should be used to determine behavior, instead of limiting the path length.

NAME_MAX depends on the file system used. Its value depends on the path, it is
not constant for the entire system.

* PR feedback

* Sync with corefx Interop.ReadLink

* Deal with /etc/localtime link containing relative path

Signed-off-by: dotnet-bot <dotnet-bot@microsoft.com>