Make sure CORINFO_FLG_FINAL is reported on delegate invoke methods
RyuJIT special cases delegate invocation and there is an assert that the special casing is safe (the method is indeed final).

We could end up not reporting `final` if we used a devirtualization algorithm that wants to prevent devirtualization there (because the delegate type was never allocated). This is pretty safe to devirtualize though (RyuJIT is not going to actually call it anyway).