Route abstract method reflectability through MetadataManager (#5145)
Progress towards #3224: route code that handles tracking reflectable abstract methods through `MetadataManager`. This is so that if we're using a `MetadataManager` that doesn't need this kind of accounting, we could even avoid allocating the `ReflectableMethodNode` nodes.

This change is also making the `ReflectableMethodNode` practically a private implementation detail of `MetadataManager`.