Give MetadataManager option to control field reflectability
The policy that all types have all of their fields reflectable was getting hardcoded in places we don't want.

I'm moving the logic down into `MetadataManager` and it's descendants so that we can have more control over this logic. This has an immediate benefit that when `EmptyMetadataManager` is in use, we no longer root all the static bases of generated types.