Normalize types and methods to fully canonical forms (#5169)
This was hit with Benchmark.net in the new code I added for reflection. We had partially canonical types and methods flying around in the system and up until the reflection work this went unnoticed. I sprinkled around some new asserts so that we'll be more likely to see a problem like this in the future.

There are a couple places where we might end up with things like `Foo<object, __Canon>`. We shouldn't let things like that float around in the system.