Fixes for virtual function resolution - Fix name/sig method search pattern in MetadataVirtualMethodAlgorithm   - Add reverse and forward scan options   - Add support for additional predicate in scan to avoid finding incorrect result - GVM support to use virtual method algorithm in its entirety instead of just the FindSlotDefiningMethodForVirtualMethod feature
Addresses issues that caused issue #190 to be brougt up. It turns out that fixing GetMethod is inappropriate to fix this sort of issue as ordering, ambiguities, and the actual set of methods that need to be considered by virtual resolution makes it inappropriate to actually use GetMethod in problematic cases.

[tfs-changeset: 1685631]