Fix MethodDesc.ToString in runtime type loader
The full `ToString` implementation includes the signature to help diagnosability, but the runtime type loader might not be able to provide that information.

Stop including MethodDesc.ToString.cs in the runtime type loader and restore old overrides on `MethodForInstantiatiatedType` and `InstantiatedMethod` that got deleted in #5065 (but only for the runtime type loader).

Also fixing up the the project file:
* The tracing options should also be available in CoreRT
* Deleting `TYPE_SYSTEM_SINGLE_THREADED` that got introduced in TFS changeset 1553623 and got never used for anything