Cleanup of remaining Toolchain tests failing in all Helix configurations
1) NegativeHighMinorVersion - according to Simon this is no longer useful; delete the test.
2) SuppressingErrorFails - the test doesn't work as it would need the Gatekeeper analysis which is known to be non-functional; delete the test.
3) RVAFields - apparently this was a negative test for a feature that now works; I have moved it to the NUTC folder and changed it into a positive test.

In 2nd iteration, I have updated the change as follows:

4) I checked in deletion of the SuppressingErrorFails test based on Vitek's review and I removed it from the presented change.
5) I have reverted the changes to NegativeHighMinorVersion as Simon has a more comprehensive review out for cleanup of these tests.
6) I have kept RVAFields and I added two more "negative turned positive" NUTC tests - MethodWithLargeFrames and OpenVirtualDelegates.

OpenVirtualDelegates is interesting as it hit what seems like a real runtime bug - in

        protected void InitializeOpenInstanceThunk(object firstParameter, IntPtr functionPointer, IntPtr functionPointerThunk)

we're initializing declaring type runtime handle to null in

            OpenMethodResolver instanceMethodResolver = new OpenMethodResolver(default(RuntimeTypeHandle), functionPointer, default(GCHandle), 0);

and that subsequently crashes hash code calculation due to an attempt to call the GetHashCode method on the null EEType*. In this iteration I have fixed this by adding a check to OpenMethodResolver but I'm not sure whether the check shouldn't be rather added directly to EETypePtr. I would be grateful if DavidWr could comment on this and I'm also adding JanK to chime in. With the additional OpenMethodResolver fix the test passes.

Thanks

Tomas

[tfs-changeset: 1688016]