Update CoreCLR tests
Picks up the latest version of Pri0 tests (as of yesterday).

There was a massive cleanup of Pri0 tests on the CoreCLR side so we now run only about 2000 tests. Hopefully, they're more representative now. Most of what we picked as "Top200" was no longer part of the Pri0 suite, so I made a new one.

We might want to consider switching to Pri1 for the rolling builds.