Move check for IsByRefLike to MakeArrayType
Having a call to `IsByRefLike` in the type unifier meant that an app that uses a simple `typeof()` would require the full reflection stack to be present in the image (all of the custom attribute resolution, method resolution, etc.).

This is a step towards having a more pay for play framework (a hello world EXE really should not have the full reflection stack compiled in it).