Avoid reflection in Enum.ToString
Contributes to #5013.

Having reflection field access and custom attribute parsing support in a code path reachable from `Enum.ToString` means that any "hello world"-style app needs to have pretty much the full reflection stack embedded in it. The reflection stack is huge.

I'm creating a shortcut EnumInfo that takes advantage of the fact that the TypeInfo is native metadata based and has a full type handle. It uses metadata APIs directly to read the metadata, bypassing the reflection stack.