Fix Path.GetTempFileName on Windows (dotnet/coreclr#17148)
GetTempFileNameW doesn't return the required length of the buffer; it returns the unique number used in the temporary file name.  But the calling code is using that result as a length, resulting in sometimes truncated paths, and sometimes paths that aren't at all related to the original due to state left over in a pooled buffer.

Signed-off-by: dotnet-bot <dotnet-bot@microsoft.com>