Extract class library composition into `.targets`
Contributes to #5662. Still not putting that library under a `Condition = Exe` because I'm not quite sure how many concepts `System.Private.DeveloperExperience.Console` is conflating (it doesn't seem like it's doing only what the name implies).

I'm doing this mostly so that the compiler is more hack-friendly for an after work project of mine.