Add option to strip exception messages from framework
This restores the feature we shipped with by default before the 2.0 release. This time under an opt-in switch.

There will be extra work needed on the framework side to make this switch actually work for all of the framework (there are things, like SR.cs, that stop working if we strip the messages right now).

This reduces the size of the People app by 412 kB, bringing us closer to something that would be acceptable to the Windows team.

When the switch is active, we'll run the exception outlining IL transform, and strip all exception messages from the executable.

Note: this is not currently hooked up E2E (not piping it to StoreAppCompile) - it's not clear if we need it there. If we do, we'll need to also pipe through the switch that disables generation of stack trace metadata because that one is a good SoD optimization too (can be done at the same time).

[tfs-changeset: 1704091]