WIP: Initial implementation of R2R section layout builder
This change introduces a new class SectionBuilder which is supposed
to accumulate compilation products to be put to the final PE file.
In accordance with my recent write-up it's based on the concept
of 'blocks' that are used to build up the PE executable sections.

A user of SectionBuilder starts out by planning out the sections;
to simplify things the section builder internally treats the sections
as zero-based integers in the order they were added to the section builder.
Their indices don't necessarily need to correspond to PE section indices
as the section indices might get shifted due to sections being merged
with the input MSIL PE file.

Once the user of the SectionBuilder class constructs the sections,
they can start adding blocks. Each block is an atomic piece of the
future output PE with alignment constraints and possibly relocations.
To simplify future tuning of this prototype I have created an abstract
BlockHandle struct to minimize churn when changing the underlying key type.

Based on my recent discussion with Vitek, I based this change on a
two-level hierarchy of "sections" and "subsections". Sections correspond
to PE sections and are externally visible in the PE file; subsections
are internal building blocks within the PE sections usable for stratifying
section content e.g. w.r.t. hot vs. cold code.

SectionBuilder is also capable of defining and producing export symbols.
As we don't yet have the complete R2R table machinery working, I hoped
to be able to demonstrate the builder by exporting a native symbol I would
manually generate as a bit of code like "mov eax, offset of a string, ret"
where the string would be a helloworld-like thing I would display in a
native caller of the DLL.

I have all of the machinery working but for some reason the Windows loader
fails to apply the file-level relocations. I have verified that the same
thing happens (the file-level relocations don't get applied) when
I LoadLibrary the primary MSIL PE file from the native host. I guess that
perhaps some header flags are used to say whether the relocations should
be applied or not.

I'm actually not clear on how the managed DLL's can work at
all as they seem to get loaded at arbitrary addresses so I guess they
need relocating all the time. For now I'm putting this off as a likely
consequence of my misunderstanding of some of the underlying concepts
as I believe the output PE file content is set up all right
according to dumpbin.

I have also created a new source file "Relocator.cs" I mostly cloned
from a bit of code I found in the CoreRT dependency analyzer; I suspect
it was most likely incomplete but it let me implement an initial
set of the most common relocators.

There's an apparent code duplication with the dependency analyzer currently,
it might be nice to find a suitable location for some "relocation utils"
after agreeing on a final version of the code.

In the pre-existing R2RPEBuilder, I addressed Michal's PR feedback from
Simon's initial change - I came to the conclusion that in this particular
case it might be cleaner to add a completely new helper struct. I converted
the remaining Tuples to named tuples per Michal's suggestion and I updated
the project file accordingly.

I have also had to introduce a new parameter to the R2RPEBuilder constructor
- directoryUpdater used to potentially adjust the PE file directories.
This is used in the SectionBuilder to inject the export directory table
and to adjust the relocation table (with a grain of salt - as I said above,
the file-level relocations still don't fully work).

For now I only have one small test case and it doesn't even pass due to
the missing relocation - in fact I can't even step the code at the DLL
location as the debugger complains about an AV; weirdly enough, the
PE-level information like section headers seems to be perfectly in order.
I still believe this might be a useful step towards an initial R2R demo.

Thanks

Tomas