Initial support for setting up the ReadyToRun header
This additional delta tweaks the R2RPEBuilder and the SectionBuilder
to allow modifying the COR header during the rewrite - this is needed
to properly publish the R2R header to the runtime.

One interesting catch is that, as the COR header points to stuff
all over the PE, it cannot be actually relocated / populated at the time
it's initially serialized into the text section.

For now I have resolved this such that section relocation only
de-serializes it and exposes it from the R2RPEBuilder class
and the subsequent SectionBuilder.EmitR2R extension does the remaining
heavy lifting in cooperation with SectionBuilder.RelocateOutputFile.

Thank

Tomas