Emit Ready-to-run header (#5988)
* Emit Ready-to-run header

Write out the ready-to-run (managed native) header, which outputs directory entries for each of the various tables. Currently we write out the CompilerIdentifierNode, a simple example showing how we'll write out the other tables.

The header is very similar to the existing CoreRT ready-to-run header with a few differences. For example, the section count is Int16 on CoreRT but Int32 on CoreCLR. The section pointers are also encoded differently (pointer relocs on CoreRT, baseless RVAs on CoreCLR). Also, each section reference can have flags on CoreRT.